refactor: main page ssr로 변경
